+++ vlm/src/test/scala/geotrellis/contrib/vlm/geotiff/GeoTiffRasterSourceSpec.scala
@@ -17,6 +17,7 @@
 package geotrellis.contrib.vlm.geotiff
 
 import geotrellis.contrib.vlm._
+import geotrellis.proj4._
 import geotrellis.raster._
 import geotrellis.raster.io.geotiff.reader.GeoTiffReader
 import geotrellis.raster.resample._
@@ -27,7 +28,6 @@ import geotrellis.spark.tiling._
 import geotrellis.util._
 import org.scalatest._
 
-import scala.concurrent.Await
 
 class GeoTiffRasterSourceSpec extends FunSpec with RasterMatchers with BetterRasterMatchers with GivenWhenThen {
   lazy val url = Resource.path("img/aspect-tiled.tif")
@@ -140,24 +140,6 @@ class GeoTiffRasterSourceSpec extends FunSpec with RasterMatchers with BetterRas
 
         layoutSource.source
       }
-    }
-
-    it("RasterSource should be threadsafe") {
-      import scala.concurrent.duration.Duration
-      import scala.concurrent.Future
-      import scala.concurrent.ExecutionContext
-
-      implicit val ec = ExecutionContext.global
-
-      val rs = GeoTiffRasterSource(url)
-
-      val res: List[Future[Option[Raster[MultibandTile]]]] = (0 to 100).toList.map { _ => Future {
-        rs.read()
-      } }
-
-      val fres: Future[List[Raster[MultibandTile]]] = Future.sequence(res).map(_.flatten)
-
-      Await.result(fres, Duration.Inf)
     }
   }
 }